+++ ddsc_site/search_indexes.py
@@ -2,22 +2,30 @@
 from __future__ import print_function, unicode_literals
 from __future__ import absolute_import, division
 
-import datetime
-from haystack.indexes import *
+import re
+from haystack import indexes
+from ddsc_core.models import Timeseries
 from ddsc_site.models import Annotation
 
-class AnnotationIndex(SearchIndex, Indexable):
-    category = CharField(model_attr='category', null=True)
-    text = CharField(document=True, use_template=False, model_attr='text', null=True)
-    username = CharField(model_attr='username', null=True)
-    picture_url = CharField(model_attr='picture_url', null=True)
-    the_model_name = CharField(model_attr='the_model_name', null=True)
-    the_model_pk = CharField(model_attr='the_model_pk', null=True)
-    location = LocationField(model_attr='location', null=True)
-    datetime_from = DateTimeField(model_attr='datetime_from', null=True)
-    datetime_until = DateTimeField(model_attr='datetime_until', null=True)
-    visibility = CharField(model_attr='visibility', null=True)
-    tags = CharField(model_attr='tags', null=True)
+
+class AnnotationIndex(indexes.SearchIndex, indexes.Indexable):
+    category = indexes.CharField(model_attr='category', null=True)
+    text = indexes.CharField(
+        document=True,
+        use_template=False,
+        model_attr='text',
+        null=True)
+    username = indexes.CharField(model_attr='username', null=True)
+    picture_url = indexes.CharField(model_attr='picture_url', null=True)
+    the_model_name = indexes.CharField(model_attr='the_model_name', null=True)
+    the_model_pk = indexes.CharField(model_attr='the_model_pk', null=True)
+    location = indexes.LocationField(model_attr='location', null=True)
+    datetime_from = indexes.DateTimeField(
+        model_attr='datetime_from', null=True)
+    datetime_until = indexes.DateTimeField(
+        model_attr='datetime_until', null=True)
+    visibility = indexes.CharField(model_attr='visibility', null=True)
+    tags = indexes.CharField(model_attr='tags', null=True)
 
     def get_model(self):
         return Annotation
@@ -27,3 +35,27 @@ class AnnotationIndex(SearchIndex, Indexable):
 
     def get_updated_field(self):
         return 'updated_at'
+
+
+class TimeseriesIndex(indexes.SearchIndex, indexes.Indexable):
+    text = indexes.CharField(document=True, use_template=True)
+    name = indexes.CharField(model_attr='name', null=True)
+    location_name = indexes.CharField(model_attr='location__name', null=True)
+    location_geom = indexes.LocationField(
+        model_attr='location__real_geometry', null=True)
+    source = indexes.CharField(model_attr='source__name', null=True)
+
+    def get_model(self):
+        return Timeseries
+
+    def index_queryset(self, using=None):
+        return self.get_model().objects.all()
+
+    def get_updated_field(self):
+        return 'updated_at'
+
+    def prepare_name(self, obj):
+        return ' '.join(re.split(r'\W+|\_', obj.name))
+
+    def prepare_location_name(self, obj):
+        return ' '.join(re.split(r'\W+|\_', obj.location.name))